Allow for blurring of background
Many adjustments have been made to CORE to allow for the implementation
of screen effects, there is a whole new function for non Sprite
effects.

There has been the addition of a blur function based off:
https://stackoverflow.com/questions/30723253/blurring-in-pygame
Which allows the background to blurred in death, while also
other effects can be added. As the method is easily scalable